Split symbolizer to style conversion into separate modules
